pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = "Python template repository"
authors = ["dogeplusplus <albertswchung@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
tensorflow = "^2.8.0"
numpy = "^1.22.3"
optax = "^0.1.1"
einops = "^0.4.1"
mlflow = "^1.24.0"
dm-haiku = "^0.0.6"
tqdm = "^4.63.1"
jax = "^0.3.4"
tensorflow-datasets = "^4.5.2"

[tool.poetry.dev-dependencies]
pytest = "^7.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
